pyproject.toml
[tool.poetry]
name = "myair"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
